+++ public/locale/en.json
@@ -3580,6 +3580,8 @@
   "no_product_definition_added": "No product definition added",
   "no_product_knowledge_found": "No product knowledges found!",
   "no_product_knowledge_found_for": "No product knowledge found for {{query}}",
+  "no_product_knowledge_linked": "No product knowledge linked, please select a product knowledge",
+  "no_product_linked": "No product linked",
   "no_product_selected": "No product selected",
   "no_products_added": "No products added",
   "no_products_found": "No products found",
@@ -5132,6 +5134,7 @@
   "select_prn_reason": "Select reason for PRN",
   "select_product": "Select Product",
   "select_product_at_row": "Please select a product at row {{row}}",
+  "select_product_first": "Select a product first",
   "select_product_knowledge": "Select product knowledge",
   "select_product_type": "Select the product type",
   "select_purchase_order": "Select Purchase Order",
